requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
rel==0.4.8
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.15.0
rpi.gpio==0.7.1
PyChromecast==13.0.4